Wiring up singleflight resolver in test cases, extracting cycle tests
